requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
flake8==3.9.2
idna==2.10
mccabe==0.6.1
mysql==0.0.3
mysql-connector-python-rf==2.2.2
mysqlclient==2.0.3
Pillow==8.3.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
stripe==2.59.0
urllib3==1.26.5
